pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["Anton Pashinov <troxin1988@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
pytils = "^0.4.1"
django-countries = "^7.6.1"
django = "4.2.9"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}


[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"
django-debug-toolbar = "^4.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
